backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Psycopg2-binary>=2.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
Pillow>=10.2.0
whitenoise>=6.6.0
gunicorn>=21.2.0
django-q2>=1.6.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
django-filter>=23.2
requests>=2.31.0
openpyxl>=3.1.2
reportlab>=4.0.0
djoser>=2.2.0
